+++ src/ZenonCli/Commands/Spork.cs
@@ -9,10 +9,10 @@ namespace ZenonCli.Commands
         public class List : ConnectionCommand
         {
             [Value(0, Default = 0, MetaName = "pageIndex")]
-            public uint? PageIndex { get; set; }
+            public int? PageIndex { get; set; }
 
             [Value(1, Default = 25, MetaName = "pageSize")]
-            public uint? PageSize { get; set; }
+            public int? PageSize { get; set; }
 
             protected override async Task ProcessAsync()
             {
@@ -25,7 +25,7 @@ namespace ZenonCli.Commands
                 AssertPageRange(PageIndex.Value, PageSize.Value);
 
                 var result = await Zdk!.Embedded.Spork
-                    .GetAll(PageIndex.Value, PageSize.Value);
+                    .GetAll((uint)PageIndex.Value, (uint)PageSize.Value);
 
                 if (result == null || result.Count == 0)
                 {
